requirements.txt
Flask
Flask-Cors
Flask-httpauth
Flask-SQLAlchemy
Flask-redis
PyMySQL
WTForms
gunicorn
captcha
tenacity
requests